# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /publications/{slug}:
    get:
      operationId: getPublication
      tags: [publications]
      summary: Get a published publication by slug
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
            minLength: 1
      responses:
        "200":
          description: Published publication
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Publication"
        "404":
          description: Publication not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Publication must be archived before permanent deletion, or changed during deletion
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /admin/publications:
    get:
      operationId: listAdminPublications
      tags: [editorial]
      summary: List all publications for administrators
      responses:
        "200":
          description: All publications ordered by most recently updated
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminPublication"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Administrator role required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    post:
      operationId: createAdminPublication
      tags: [editorial]
      summary: Create a publication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublicationInput"
      responses:
        "201":
          description: Publication created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminPublication"
        "400":
          description: Invalid publication
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Administrator role required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Publication slug already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /admin/publications/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
    get:
      operationId: getAdminPublication
      tags: [editorial]
      summary: Get a publication by ID
      responses:
        "200":
          description: Publication
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminPublication"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Administrator role required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Publication not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    patch:
      operationId: updateAdminPublication
      tags: [editorial]
      summary: Update, publish, or archive a publication by ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublicationUpdate"
      responses:
        "200":
          description: Updated publication
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminPublication"
        "400":
          description: Invalid publication update
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Administrator role required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Publication not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Publication slug already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteAdminPublication
      tags: [editorial]
      summary: Permanently delete a publication by ID
      responses:
        "204":
          description: Publication deleted
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Administrator role required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Publication not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    PublicationAuthor:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        role:
          type: string
    PublicationType:
      type: string
      enum:
        - research-paper
        - policy-brief
        - article
        - commentary
        - report
        - case-study
    PublicationStatus:
      type: string
      enum:
        - draft
        - published
        - archived
    Publication:
      type: object
      required:
        - id
        - slug
        - title
        - abstract
        - publicationType
        - authors
        - publicationDate
        - featured
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        title:
          type: string
        subtitle:
          type: string
        abstract:
          type: string
        publicationType:
          $ref: "#/components/schemas/PublicationType"
        category:
          type: string
        themeId:
          type: ["string", "null"]
          format: uuid
        language:
          type: string
          enum: [english, french, bilingual]
        authors:
          type: array
          items:
            $ref: "#/components/schemas/PublicationAuthor"
        publicationDate:
          type: string
          format: date
        readingTime:
          type: integer
          minimum: 1
        featured:
          type: boolean
        featuredImage:
          type: string
        featuredImageMediaId:
          type: ["string", "null"]
          format: uuid
        pdfUrl:
          type: string
        pdfMediaId:
          type: ["string", "null"]
          format: uuid
        externalUrl:
          type: string
        doi:
          type: string
        content:
          type: string
        seoTitle:
          type: string
        seoDescription:
          type: string
    AdminPublication:
      type: object
      required:
        - id
        - slug
        - title
        - abstract
        - publicationType
        - authors
        - featured
         - themeId
         - language
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        title:
          type: string
        subtitle:
          type: string
        abstract:
          type: string
        publicationType:
          $ref: "#/components/schemas/PublicationType"
        category:
          type: string
        themeId:
          type: ["string", "null"]
          format: uuid
        language:
          type: string
          enum: [english, french, bilingual]
        authors:
          type: array
          items:
            $ref: "#/components/schemas/PublicationAuthor"
        publicationDate:
          type: string
          format: date
        readingTime:
          type: integer
          minimum: 1
        featured:
          type: boolean
        featuredImage:
          type: string
        featuredImageMediaId:
          type: ["string", "null"]
          format: uuid
        pdfUrl:
          type: string
        pdfMediaId:
          type: ["string", "null"]
          format: uuid
        externalUrl:
          type: string
        doi:
          type: string
        content:
          type: string
        seoTitle:
          type: string
        seoDescription:
          type: string
        status:
          $ref: "#/components/schemas/PublicationStatus"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        createdBy:
          type: ["string", "null"]
        updatedBy:
          type: ["string", "null"]
        publishedBy:
          type: ["string", "null"]
        publishedAt:
          type: ["string", "null"]
          format: date-time
        archivedBy:
          type: ["string", "null"]
        archivedAt:
          type: ["string", "null"]
          format: date-time
    PublicationInput:
      type: object
      required:
        - slug
        - title
        - abstract
        - publicationType
        - authors
        - featured
        - status
      properties:
        slug:
          type: string
          minLength: 1
          pattern: "^[a-z0-9-]+$"
        title:
          type: string
          minLength: 1
        subtitle:
          type: string
          nullable: true
        abstract:
          type: string
          minLength: 1
        publicationType:
          $ref: "#/components/schemas/PublicationType"
        category:
          type: string
          nullable: true
        themeId:
          type: string
          format: uuid
          nullable: true
        language:
          type: string
          enum: [english, french, bilingual]
        authors:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/PublicationAuthor"
        publicationDate:
          type: string
          format: date
          nullable: true
        readingTime:
          type: integer
          minimum: 1
          nullable: true
        featured:
          type: boolean
        featuredImage:
          type: string
          format: uri
          nullable: true
        featuredImageMediaId:
          type: string
          format: uuid
          nullable: true
        pdfUrl:
          type: string
          format: uri
          nullable: true
        pdfMediaId:
          type: string
          format: uuid
          nullable: true
        externalUrl:
          type: string
          format: uri
          nullable: true
        doi:
          type: string
          nullable: true
        content:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/PublicationStatus"
        seoTitle:
          type: string
          nullable: true
        seoDescription:
          type: string
          nullable: true
    PublicationUpdate:
      minProperties: 1
      type: object
      properties:
        slug:
          type: string
          minLength: 1
          pattern: "^[a-z0-9-]+$"
        title:
          type: string
          minLength: 1
        subtitle:
          type: string
          nullable: true
        abstract:
          type: string
          minLength: 1
        publicationType:
          $ref: "#/components/schemas/PublicationType"
        category:
          type: string
          nullable: true
        themeId:
          type: string
          format: uuid
          nullable: true
        language:
          type: string
          enum: [english, french, bilingual]
        authors:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/PublicationAuthor"
        publicationDate:
          type: string
          format: date
          nullable: true
        readingTime:
          type: integer
          minimum: 1
          nullable: true
        featured:
          type: boolean
        featuredImage:
          type: string
          format: uri
          nullable: true
        featuredImageMediaId:
          type: string
          format: uuid
          nullable: true
        pdfUrl:
          type: string
          format: uri
          nullable: true
        pdfMediaId:
          type: string
          format: uuid
          nullable: true
        externalUrl:
          type: string
          format: uri
          nullable: true
        doi:
          type: string
          nullable: true
        content:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/PublicationStatus"
        seoTitle:
          type: string
          nullable: true
        seoDescription:
          type: string
          nullable: true
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: string
...