# lib/api-zod/src/generated/types/adminPublication.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AdminPublicationLanguage } from './adminPublicationLanguage';
import type { PublicationAuthor } from './publicationAuthor';
import type { PublicationStatus } from './publicationStatus';
import type { PublicationType } from './publicationType';

export interface AdminPublication {
  id: string;
  slug: string;
  title: string;
  subtitle?: string;
  abstract: string;
  publicationType: PublicationType;
  category?: string;
  /** @nullable */
  themeId?: string | null;
  language?: AdminPublicationLanguage;
  authors: PublicationAuthor[];
  publicationDate?: string;
  /** @minimum 1 */
  readingTime?: number;
  featured?: boolean;
  featuredImage?: string;
  /** @nullable */
  featuredImageMediaId?: string | null;
  pdfUrl?: string;
  /** @nullable */
  pdfMediaId?: string | null;
  externalUrl?: string;
  doi?: string;
  content?: string;
  seoTitle?: string;
  seoDescription?: string;
  status: PublicationStatus;
  createdAt: string;
  updatedAt: string;
  /** @nullable */
  createdBy?: string | null;
  /** @nullable */
  updatedBy?: string | null;
  /** @nullable */
  publishedBy?: string | null;
  /** @nullable */
  publishedAt?: string | null;
  /** @nullable */
  archivedBy?: string | null;
  /** @nullable */
  archivedAt?: string | null;
}