# lib/api-zod/src/generated/types/publicationStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type PublicationStatus = typeof PublicationStatus[keyof typeof PublicationStatus];


export const PublicationStatus = {
  draft: 'draft',
  published: 'published',
  archived: 'archived',
} as const;